Split benchmarks into separate executables
